Convert EuiListGroup to TypeScript
