+++ src/components/list_group/index.ts
@@ -1,3 +1,2 @@
 export { EuiListGroup } from './list_group';
-
 export { EuiListGroupItem } from './list_group_item';
